Json output added
Dear Gregor,

as requested I have added the json file output to the mmb cli. This is based on the file I have sent you before.
NOTE: the file MMBOPT1.m I sent on the 21st of June, is a for the GUI, it was meant to test the json on a platform Jan was familiar with. Going to the cli MMB file is a bit more involving, especially for option 2 as there the structure changes. Now it works for both, but the Jsonstructure is not "in the schema" Jan sent.

I have tried to implement the schema but I did not manage, please consult line 577 for the try. I believe the best is to loop through the fields of the structure in json to create a table from it. That can be then stored in the format Jan requested using matlab's built in funtion: jsonencode