Json structure added for option1- i.e. it will run with one model one type of rule and one shock
Dear all,
please find the updates to the CMD_MMB.m file and the respective output commited. It works with the one model one rule, one shock, ie the basic case. It is a good start to add it as a functionality for output that can only be produced using option2 (eg model specific shocks, combined shocks).
Please review it and use it to develop the functionality further!